refactor: moved sort list data to product list domain
