+++ src/app/(store)/list/[[...params]]/page.tsx
@@ -12,9 +12,8 @@ import NoItem from "@/components/store/listPage/noItem";
 import Button from "@/components/UI/button";
 import DropDownList from "@/components/UI/dropDown";
 import LineList from "@/components/UI/lineList";
-import { sortDropdownData } from "@/data/uiElementsData";
 import { ProductListSkeleton } from "@/domains/store/productList/components";
-import { DEFAULT_FILTERS, SORT_DATA } from "@/domains/store/productList/constants";
+import { DEFAULT_FILTERS, SORT_DATA, sortDropdownData } from "@/domains/store/productList/constants";
 import { TFilterBrands, TFilters, TListItem } from "@/domains/store/productList/types";
 import { getFiltersFromProductList } from "@/domains/store/productList/utils";
 import { IMAGE_BASE_URL } from "@/shared/constants/store";
