--- src/data/uiElementsData.ts
@@ -1,24 +1,0 @@
-import { TDropDown } from "@/types/uiElements";
-
-export const sortDropdownData: TDropDown[] = [
-  {
-    text: "Newest",
-    value: "0",
-  },
-  {
-    text: "Oldest",
-    value: "1",
-  },
-  {
-    text: "Most Expensive",
-    value: "2",
-  },
-  {
-    text: "Cheapest",
-    value: "3",
-  },
-  {
-    text: "Name",
-    value: "4",
-  },
-];
